Adds restrictions to bed interaction + dialogue prompts
Bed cannot be interacted with before the notepad is picked up. If it is, dialogue pops up prompting the user to pick the notepad.

Adds a dialogueUIUpdater in the scene that adds dialogue message prompts that persist for 5 seconds. Dialogue messages appear at the bottom of the screen and tutorial text appears at the top of the screen

Also makes changes to tutorial text. The first dialogue msg appears in a different area than the tutorial text.

Adds dialogue popup when the player tries to open the door before completing the tutorial